Remoove requirement that vectors have same length

+++ PhysicsTools/NanoAOD/python/muons_cff.py
@@ -234,7 +234,7 @@ muonExternalVecVarsTable = cms.EDProducer("FlattenedCandValueMapVectorTableProdu
     doc = muonTable.doc,
     variables = cms.PSet(
         # can you declare a max number of bits here?  technically for the moment this needs 16 bits, but might eventually need 17 or 18
-        GlobalIdxs = ExtVar(cms.InputTag("trackrefit:globalIdxs"), "std::vector<int>", doc="Indices for correction parameters", precision = 16),
+        GlobalIdxs = ExtVar(cms.InputTag("trackrefit:globalIdxs"), "std::vector<int>", doc="Indices for correction parameters"),
         # optimal precision tbd, but presumably can work the same way as for scalar floats
         JacRef = ExtVar(cms.InputTag("trackrefit:jacRef"), "std::vector<float>", doc="jacobian for corrections", precision = 12),
         # optimal precision tbd, but presumably can work the same way as for scalar floats
